requirements.txt
torch>=2.4.1
pandas>=2.2.3
transformers>=4.45.1
prettytable>=3.16.0
tqdm>=4.66.5
pillow>=10.4.0
torchvision>=0.19.1
numpy>=2.0.1
janus>=1.0.0